+++ src/vue/common/AccountBalances.vue
@@ -1,13 +1,17 @@
 <template>
   <div class="account-balances">
     <template v-if="balancesAll && isLoadedStaking">
-      <p class="account-balances__total">
+      <p
+        v-tooltip="$fFullBalance(currentBalance)"
+        class="account-balances__total"
+      >
         {{ $fbalance(totalBalance) }}
       </p>
 
       <readonly-row
         class="account-balances__row"
         v-if="!availableBalance.isZero()"
+        v-tooltip="$fFullBalance(availableBalance)"
         :label="$t('common.account-balances.transferrable-lbl')"
         :value="$fbalance(availableBalance)"
       />
@@ -15,6 +19,7 @@
       <readonly-row
         class="account-balances__row"
         v-if="!balancesAll.lockedBalance.isZero()"
+        v-tooltip="$fFullBalance(balancesAll.lockedBalance)"
         :label="$t('common.account-balances.locked-lbl')"
         :value="$fbalance(balancesAll.lockedBalance)"
       />
@@ -22,6 +27,7 @@
       <readonly-row
         class="account-balances__row"
         v-if="+stakingBalance.balance && !isLoadStakingFailed"
+        v-tooltip="$fFullBalance(stakingBalance.balance)"
         :label="$t('common.account-balances.staked-lbl')"
         :value="$fbalance(stakingBalance.balance)"
       />