+++ src/vue/forms/transfer-form/TransferSendAuthorizeStep.vue
@@ -8,13 +8,19 @@
         <h3 class="transfer-form-authorize-step__title">
           {{ $t('forms.transfer-form-authorize-step.amount-title') }}
         </h3>
-        <p class="transfer-form-authorize-step__amount-value">
+        <p
+          v-tooltip="$fFullBalance(former.attrs.amount)"
+          class="transfer-form-authorize-step__amount-value"
+        >
           {{ $fbalance(former.attrs.amount) }}
         </p>
         <h3 class="transfer-form-authorize-step__title">
           {{ $t('forms.transfer-form-authorize-step.submission-fee-title') }}
         </h3>
-        <p class="transfer-form-authorize-step__fee-value">
+        <p
+          v-tooltip="$fFullBalance(partialFee)"
+          class="transfer-form-authorize-step__fee-value"
+        >
           {{ $fbalance(partialFee) }}
         </p>