+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - No data message on staking tab
 - Open graph meta information
 - Data to connect the network in metamask
+- Full balance tooltip
 
 #### Changed
 - Version api packages